Rewrite do_pending_operator as a pipeline and a named switch
The batch's widest switch, and the last transpiled function in the
family. 697 lines become a driver plus nine steps, and the module doc is
a table of what each one decides -- because the hard part of this
function is not the switch, it is the four hundred lines before it that
work out what the region *is*:

    apply_motion_force      v/V/CTRL-V typed between operator and motion
    record_operator_redo    what `.` replays, and what is not replayable
    resume_redo_visual      `.` replaying: the size comes from REDO_VISUAL
    start_visual_region     the operator was typed after a selection
    order_region            which end is the start; closed folds
    prepare_visual_redo     the selection's size, for `gv` and the next `.`
    finish_visual_region    line/charwise fixups; Visual off
    adjust_region_end       an exclusive end in column one
    run_operator            the switch

**The switch now spells the operators.** It was `match (*oap).op_type {
4 | 5 => ..., 14 | 13 => ..., 26 => ... }` over thirty bare numbers;
every arm is an `OP_*` name, which is what `OP_JOIN_NS` and `OP_FORMAT2`
were declared for. `OP_FILTER`'s `FALLTHROUGH` into the `=`/`:` arm is
an explicit call to the shared `indent_or_colon`, and the two arms that
enter Insert mode are `run_change` and `run_block_insert`.

Two locals become named things: the function-static `redo_VIsual` is
`REDO_VISUAL` at module scope with a doc comment saying why it is
process-wide (`.` after `viwd` deletes the same *number* of characters),
and the seven-way `op_type != OP_FOLD*` chain -- written out twice -- is
a `matches!`.

`clear_oparg`'s `memset` is `*oap = oparg_T::ZERO`.

One real bug, caught by `oldtest test_visual` and by nothing else:
`op_delete`'s blockwise arm sizes its allocation `ml_get_len(lnum) -
n + 1` where `n` can be **negative** (deleting a partly-covered TAB
replaces it with more spaces than it took), and upstream does that
arithmetic in `size_t` and relies on the wraparound. Written as a
`size_t` subtraction it panicked in a debug build; it is `c_int`
arithmetic now, with the reason at the site. The `traps-code` entry to
write down: **a `wrapping_sub` c2rust emitted on a `size_t` is
load-bearing whenever the operand can be negative**, and opsweep does
not reach it -- only `Test_visual_block_mode` does.

`--allow-growth`: `unsafe ` 7 -> 20 over ten new functions. clippy 2 ->
0. opsweep byte-identical; `oldtest test_normal test_visual test_put
test_fold test_edit test_increment test_registers test_ex_mode` all
green (485 tests).

+++ crates/nvim/src/nvim/ops/delete.rs
@@ -210,10 +210,14 @@ unsafe fn delete_block(oap: *mut oparg_T) -> c_int {
                 }
 
                 // The line shrinks by the block's text minus the padding that
-                // replaces the characters it only partly covers.
+                // replaces the characters it only partly covers -- and a
+                // deleted TAB can be replaced by more spaces than it took, so
+                // `n` may be *negative* and the line grow. The arithmetic
+                // stays in `c_int` for that reason; upstream does it in
+                // `size_t` and relies on the wraparound.
                 let n = bd.textlen - bd.startspaces - bd.endspaces;
                 let oldp = ml_get(lnum);
-                let newp = xmalloc(ml_get_len(lnum) as size_t - n as size_t + 1) as *mut c_char;
+                let newp = xmalloc((ml_get_len(lnum) - n + 1) as size_t) as *mut c_char;
                 memmove(
                     newp as *mut c_void,
                     oldp as *const c_void,
